+++ tests/test_rest.py
@@ -11,8 +11,6 @@ import tempfile
 import os
 from copy import deepcopy # copy test config
 
-import time
-
 import pytest
 from fastapi.testclient import TestClient
 
@@ -985,56 +983,16 @@ def test_generate(client):
   #   if os.path.exists('{}/{}'.format(fullpath, fn)):
   #     os.remove('{}/{}'.format(fullpath, fn))
 
-def test_zeroconf_deprecated():
-  """ Unit test for older, deprecated zeroconf advertisement """
-  # TODO: migrate this test into its own module
-  from zeroconf import Zeroconf, ServiceStateChange, ServiceBrowser, IPVersion
-  from time import sleep
-  from multiprocessing import Process, Queue
-
-  AMPLIPI_ZC_NAME = 'amplipi-api._http._tcp.local.'
-
-  services_advertised = {}
-  def on_service_state_change(zeroconf: Zeroconf, service_type: str, name: str, state_change: ServiceStateChange):
-    if state_change is ServiceStateChange.Added:
-      info = zeroconf.get_service_info(service_type, name)
-      if info and info.port != 80: # ignore the actual amplipi service on your network
-        services_advertised[name] = info
-
-  # advertise amplipi-api service (start this before the listener to verify it can be found after advertisement)
-  q = Queue()
-  zc_reg = Process(target=amplipi.app.advertise_service, args=(9898,q))
-  zc_reg.start()
-  sleep(4) # wait for a bit to make sure the service is started
-
-  # start listener that adds available services
-  zeroconf = Zeroconf(ip_version=IPVersion.V4Only)
-  services = ["_http._tcp.local."]
-  _ = ServiceBrowser(zeroconf, services, handlers=[on_service_state_change])
-
-  # wait enough time for a response from the serice
-  sleep(2)
-
-  # stop the advertiser
-  q.put('done')
-  zc_reg.join()
-
-  # stop the listener
-  zeroconf.close()
-
-  # check advertisememts
-  assert AMPLIPI_ZC_NAME in services_advertised
-  assert services_advertised[AMPLIPI_ZC_NAME].port == 9898
-
 def test_zeroconf():
   """ Unit test for zeroconf advertisement """
   # TODO: migrate this test into its own module
   from zeroconf import Zeroconf, ServiceStateChange, ServiceBrowser, IPVersion
   from time import sleep
-  from multiprocessing import Process, Queue
+  from multiprocessing import Process, Event
 
   # get default network interface
   iface = ni.gateways()['default'][ni.AF_INET][1]
+  FAKE_PORT = 9898
 
   # first time ni.ifaddresses is called in the CI system it fails
   try:
@@ -1051,8 +1009,8 @@ def test_zeroconf():
         services_advertised[name] = info
 
   # advertise amplipi-api service (start this before the listener to verify it can be found after advertisement)
-  q = Queue()
-  zc_reg = Process(target=amplipi.app.advertise_service, args=(9898,q))
+  event = Event()
+  zc_reg = Process(target=amplipi.app.advertise_service, args=(FAKE_PORT,event))
   zc_reg.start()
   sleep(4) # wait for a bit to make sure the service is started
 
@@ -1074,7 +1032,7 @@ def test_zeroconf():
   AMPLIPI_ZC_NAME = f'amplipi-{mac_addr}._amplipi._tcp.local.'
 
   # stop the advertiser
-  q.put('done')
+  event.set()
   zc_reg.join()
 
   # stop the listener
@@ -1082,7 +1040,7 @@ def test_zeroconf():
 
   # check advertisememts
   assert AMPLIPI_ZC_NAME in services_advertised
-  assert services_advertised[AMPLIPI_ZC_NAME].port == 9898
+  assert services_advertised[AMPLIPI_ZC_NAME].port == FAKE_PORT
 
 @pytest.mark.parametrize('zid', base_zone_ids())
 def test_set_zone_vol(client, zid):